Long double used instead of double to store first power and changes to prevent rounding problems in calculation of first power. Log logic calls added to add/calculate method.
Bug fix in Phy Rx Carrier: Interference event stored to memebr variable only when packet is received.

+++ model/satellite-per-packet-interference.h
@@ -119,7 +119,7 @@ private:
   /**
    *
    */
-  typedef std::pair <uint32_t, double > InterferenceChange;
+  typedef std::pair <uint32_t, long double > InterferenceChange;
 
   /**
    *
@@ -152,7 +152,7 @@ private:
   /**
    * \brief First power value for interference. Sum of negative values in list m_changes, which positive value is not in list
    */
-  double m_firstPowerW;
+  long double m_firstPowerW;
 
   /**
    * \brief flag to indicate that at least one receiving is on
